Allow configuring the separator for window status icons
The default space separator in front of the status icons can now be
removed, or changed to something else.

Also, the separator logic is now enhanced; the separator is only added
if the status icon is not empty. Unfortunately, it does not work with
status as text, because the value is not available yet when the
separator is added.