fix(weather status): Use tmux-weather's cache var
xamut/tmux-weather looks for '#{weather}' in status-right to replace
with the downloaded weather data.
However it can't find it as status-right will only contain
'#{E:@catppuccin_status_weather}'

So instead of using the literal '#{weather}', use the var that
xamut/tmux-wether uses to cache it's results.

Tested on my own systems.