docs: Correct the plugin path for TPM in the readme (#492)
TPM recommends the directory ~/.tmux/plugins/tpm be used as the installation directory in the README, which is likely what most users will choose. TPM can be installed into any directory, but the readme should ideally mention the most common choice that users will have, which we can assume is going to be ~/.tmux/plugins.