docs: stop recommending `-o` (#524)
Remove -o from documentation and add warnings to source files

Recurring issues are being raised that options are not being applied,
where users have used '-o' to set options. '-o' instructs tmux
to only set the option if it is not already set. This is useful
in the plugin itself, but almost never useful for the end user.